// src/types/trial.ts
export interface SpecificResponse {
  allegation_number: number;
  allegation_summary: string;
  response: string;
  admission: "admitted" | "denied" | "not_known";
}